Import RatingAverage into Product and Rating cs

+++ BackEnd/Models/Rating.cs
@@ -3,7 +3,7 @@ namespace BackEnd.Models
     public class Rating
     {
         public int RatingID { get; set; }
-        public int RatingText { get; set; }
+        public double RatingText { get; set; }
         public int ProductID  {get ; set;}
         public Product Product { get; set; }
         public string UserName { get; set; }
