+++ Share/RatingVm.cs
@@ -3,8 +3,10 @@ namespace Share
     public class RatingVm
     {
         public int Id{get;set;}
-        public int RatingText { get; set; }
+        public double RatingText { get; set; }
         public string UserName { get; set; }
         public int ProductID { get; set; }
+
+        
     }
 }
